moles
Becker's nevus is a benign skin disorder that causes a dark, irregular pigmentation and excess hair growth on the skin
Melanocytic nevi, also known as moles, are non-cancerous growths on the skin that are made up of pigment-producing cells. They can appear anywhere on the body and are usually round or oval-shaped.
Atypical nevi, also known as dysplastic nevi, are moles that look different from common moles. They are benign, but they can indicate a higher risk of developing melanoma.